Increase test coverage
